PSP #2

+++ Server/Student/controllerStudent.js
@@ -1,3 +1,4 @@
+const circle = require('../Circle/controllerCircle');
 const changeMail = require('./changeMailMail');
 const constants = require('./constants');
 const database = require('./database');
@@ -7,6 +8,8 @@ const registration = require('./registration');
 const resetPwd = require('./passwordResetMail');
 const mySession = require('../Session/session');
 const tests = require('./tests');
+const responder = require('./responseSender');
+
 
 module.exports = {
 
@@ -26,25 +29,20 @@ module.exports = {
         var validationKey = req.params.validationKey;
 
         if (!validationKey) {
-            res.status(400);
-            res.send("Bad request. No uuid.");
+            responder.sendResponse(res, 400, "Bad request. No uuid.");
             return;
         }
-
         try {
             if ( await database.validationKeyExists(validationKey)) {
                 console.log("validation key exists");
                 if (await database.setState(validationKey, constants.AccountState.ACTIVE)){
-                    res.status(201);
-                    res.send("Successfully validated new user account.");
+                    responder.sendResponse(res, 201, "Successfully validated new user account.");
                 }
             } else {
-                res.status(401);
-                res.send("Unauthorized. Invalid validation key.");
+                responder.sendResponse(res, 401, "Unauthorized. Invalid validation key.");
             }
         } catch (err) {
-            res.status(500);
-            res.send("Server Error");
+            responder.sendResponse(res, 500);
         }
     },
 
@@ -53,8 +51,7 @@ module.exports = {
         var mail = req.body.mail;
 
         if (!mail) {
-            res.status(400);
-            res.send("Bad request. No mail.");
+            responder.sendResponse(res, 400, "Bad request. No mail.");
             return;
         }
 
@@ -62,21 +59,17 @@ module.exports = {
             if (await database.userMailExists(mail)) {
                 await resetPwd.reset(mail);
             }
-            res.status(200);
-            res.send("Reset mail sent if user is known.");
+            responder.sendResponse(res, 200, "Reset mail sent if user is known.");
         } catch (err) {
             console.log(err);
-            res.status(500);
-            res.send("Server Error");
+            responder.sendResponse(res, 500);
         }
     },
 
     passwordResetPage : function (req, res) {
-        var validationKey = req.params.validationKey;
+        let validationKey = req.params.validationKey;
         if (!validationKey) {
-            res.status(400);
-            res.send("Bad request. No validation key");
-            return;
+            responder.sendResponse(res, 400, "Bad request. No validation key.");
         } else {
             res.status(200);
             res.send(passwordResetForm.getForm(validationKey));
@@ -85,48 +78,48 @@ module.exports = {
 
     //Called when user sends a new Password after requesting a password reset mail using the client
     resetPassword : async function (req, res) {
-        var validationKey = req.params.validationKey;
-        var newPassword = req.body.pwd;
+        let validationKey = req.body.validationKey;
+        let newPassword = req.body.pwd;
 
         console.log(validationKey + " | " + newPassword);
 
         if (!validationKey || !newPassword) {
-            res.status(400);
-            res.send("Bad request. No validation key or password.");
+            responder.sendResponse(res, 400, "Bad request. No validation key or password.");
             return;
         }
 
         try {
             if (await database.validationKeyExists(validationKey)) {
-                var userId = await database.getUserIdFromValidationKey(validationKey);
+                try {
+                    var userId = await database.getUserIdFromValidationKey(validationKey);
 
-                var userAuthData = passwordUtil.generateUserAuthData(newPassword);
-                var hash = userAuthData.hash;
-                var salt = userAuthData.salt;
+                    var userAuthData = passwordUtil.generateUserAuthData(newPassword);
+                    var hash = userAuthData.hash;
+                    var salt = userAuthData.salt;
 
-                await database.setPassword(userId, hash, salt);
+                    await database.setPassword(userId, hash, salt);
 
-                res.status(200);
-                res.send("Password successfully reset.");
+                    responder.sendResponse(res, 200, "Password successfully reset.");
+                } catch (err) {
+                    console.log(err);
+                    responder.sendResponse(res, 500);
+                }
             } else {
-                res.status(401);
-                res.send("Invalid validation key!");
+                responder.sendResponse(res, 401, "Invalid validation key!");
             }
         } catch (err) {
             console.log(err);
-            res.status(500);
-            res.send("Server Error");
+            responder.sendResponse(res, 500);
         }
     },
 
     //Called when the user wants to login to studiCircle. Will send session an d user data if credentials are valid
     login : async function (req, res) {
-        var mail = req.body.mail;
-        var pass = req.body.pwd;
+        let mail = req.body.mail;
+        let pass = req.body.pwd;
 
         if (!mail || !pass) {
-            res.status(400);
-            res.send("Bad request. Either no username or no password.");
+            responder.sendResponse(res, 400, "Bad request. Either no username or no password.");
             return;
         }
         try {
@@ -141,7 +134,7 @@ module.exports = {
 
                 if (passwordUtil.passwordCorrect(pass, userAuthData.salt, userAuthData.hash)) {
                     var returnObject = {};
-                    returnObject.status = 200;
+                    returnObject.httpStatus = 200;
                     returnObject.message = "Successfully Logged in";
                     returnObject.userData = (await database.getUserData(userId));
                     returnObject.session = mySession.generateSession(userId);
@@ -149,24 +142,21 @@ module.exports = {
                     res.status(200);
                     res.send(returnObject);
                 } else {
-                    res.status(401);
-                    res.send('Unauthorized! Wrong Password');
+                    responder.sendResponse( res, 401, "Unauthorized! Wrong Password");
                 }
             } else {
-                res.status(412);
-                res.send("Profile not activated!");
+                responder.sendResponse(res, 412, "Profile not activated!");
             }
         } catch (err) {
-        console.log(err)
-        res.status(500);
-        res.send("Server Error");
+            console.log(err);
+            responder.sendResponse(res, 500);
         }
     },
 
     logout : function (req, res) {
         mySession.invalidate(req.session.sessionId);
 
-        res.send("Logout successfull.")
+        responder.sendResponse(res, 200,"Logout successfull.");
     },
 
     //Called when the user sets a new password
@@ -177,8 +167,7 @@ module.exports = {
         var newPw = req.body.newPwd;
 
         if (!userId || !oldPw || !newPw || !passwordUtil.passwordIsCompliant(newPw)) {
-            res.status(400);
-            res.send("Bad request. No session, old or new password not set or not compliant to guidelines.");
+            responder.sendResponse(res, 400, "Bad request. No session, old or new password not set or not compliant to guidelines.");
             return;
         }
 
@@ -191,14 +180,13 @@ module.exports = {
                 await database.setPassword(userId, newUserAuthData.hash, newUserAuthData.salt);
 
                 res.status(200);
-                res.send("Successfully set Password")
+                res.send(true);
+//                responder.sendResponse(res, 200, "Successfully set Password");
             } else {
-                res.status(401);
-                res.send("Unauthorized. Invalid password!")
+                responder.sendResponse(res, 401, "Unauthorized. Invalid password!");
             }
         } catch (err) {
-            res.status(500);
-            res.send("Server Error");
+            responder.sendResponse(res, 500);
         }
     },
 
@@ -208,8 +196,7 @@ module.exports = {
         var pass = req.body.pwd;
 
         if (!userId || !pass) {
-            res.status(400);
-            res.send("Bad request. Either no session or no password.");
+            responder.sendResponse(res, 400, "Bad request. Either no session or no password.");
             return;
         }
 
@@ -217,45 +204,50 @@ module.exports = {
             var userAuthData = await database.getUserAuthData(userId);
 
             if (passwordUtil.passwordCorrect(pass, userAuthData.salt, userAuthData.hash)) {
-                await database.deleteUser(userId);
-                res.status(200);
-                res.send("Successfully deleted Account");
-                req.session.reset();
+
+                 circle.isAdminAnywhere(userId, async function(userIsAdmin) {
+
+                     if (!userIsAdmin) {
+                         await database.deleteUser(userId);
+                         responder.sendResponse(res, 200, "Successfully deleted Account");
+                         req.session.reset();
+                     } else {
+                         responder.sendResponse(res, 412, "User still Admin in one or more circles");
+                     }
+                 });
+
             } else {
-                res.status(401);
-                res.send("Unauthorized. Invalid password.")
+                responder.sendResponse(res, 401, "Unauthorized. Invalid password.");
             }
 
         } catch (err) {
-            console.log(err)
-            res.status(500);
-            res.send("Server Error");
+            console.log(err);
+            responder.sendResponse(res, 500);
         }
     },
 
     //
     updateMail : async function (req, res) {
+        console.log("update Mail");
         var userId = req.session.userId;
         var oldMail = req.body.oldMail;
         var newMail = req.body.newMail;
         var pass = req.body.pwd;
 
+        console.log("update Mail");
         if (!userId || !oldMail || !newMail || !pass) {
-            res.status(400);
-            res.send("Bad request. Either no session, oldMail, newMail or no password.");
+            responder.sendResponse(res, 400, "Bad request. Either no session, oldMail, newMail or no password.");
             return;
         }
 
         if (oldMail == newMail) {
-            res.status(400);
-            res.send("Bad request. Old mail is new mail.");
+            responder.sendResponse(res, 400, "Bad request. Old mail is new mail.");
             return;
         }
 
         try {
             if (!userId == (await database.getUserIdFromMail(oldMail))) {
-                res.status(401);
-                res.send("Unauthorized! Mail and session do not match!");
+                responder.sendResponse(res, 401, "Unauthorized! Mail and session do not match!");
                 return;
             }
 
@@ -263,18 +255,14 @@ module.exports = {
 
             if (passwordUtil.passwordCorrect(pass, userAuthData.salt, userAuthData.hash)) {
                 changeMail.send(oldMail, newMail);
-                res.status(200);
-                res.send("Send an validation E-Mail");
-                return;
+                responder.sendResponse(res, 200, "Send an validation E-Mail");
             } else {
-                res.status(401);
-                res.send("Error updating the email")
+                responder.sendResponse(res, 401, "Error updating the email");
             }
 
         } catch (err) {
-            console.log(err)
-            res.status(500);
-            res.send("Server Error");
+            console.log(err);
+            responder.sendResponse(res, 500);
         }
 
     },
@@ -283,8 +271,7 @@ module.exports = {
         var validationKey = req.params.validationKey;
 
         if (!validationKey) {
-            res.status(400);
-            res.send("Bad request. No validation key.");
+            responder.sendResponse(res, 400, "Bad request. No validation key.");
             return;
         }
 
@@ -295,17 +282,13 @@ module.exports = {
 
                 await database.updateMail(userId, newMail);
 
-                res.status(200);
-                res.send("Successfully updated mail address");
+                responder.sendResponse(res, 200, "Successfully updated mail address");
             } else {
-                res.status(401);
-                res.send("Unauthorized. Invalid validation key.");
-                return;
+                responder.sendResponse(res, 401, "Unauthorized. Invalid validation key.");
             }
         } catch (err) {
-            console.log(err)
-            res.status(500);
-            res.send("Server Error");
+            console.log(err);
+            responder.sendResponse(res, 500);
         }
     },
 
@@ -314,12 +297,12 @@ module.exports = {
 
         try {
             console.log("start");
-            var result = await database.setPassword(13, "hash", "salt");
+            let result = await database.setPassword(13, "hash", "salt");
             console.log(result);
         } catch (err) {
             console.log("error at set password: " + err);
         }
-        res.send('OK');
+        responder.sendResponse(res, 200, "OK");
     },
 
     test : function (req, res) {
@@ -327,7 +310,6 @@ module.exports = {
     },
 
     unknownpage : function (req, res) {
-      res.status(404);
-      res.send('Unknown Endpoint')
+        responder.sendResponse(res, 404, "Unknown Endpoint");
     }
 };
